delete page

+++ session_36/app/main/routes.py
@@ -1,5 +1,7 @@
+from werkzeug.utils import redirect
+
 from app.main import main
-from flask import render_template, request, session
+from flask import render_template, request, session, url_for
 from app.main.models import Event
 from app.extensions import db
 
@@ -32,14 +34,23 @@ def add_event():
     return render_template('add_event.html', message=message)
 
 
-@main.route('/delete_event')
+@main.route('/delete_event', methods=['GET', 'POST'])
 def delete_event():
-    hardcoded_id = 1
-    event = Event.query.filter_by(id=hardcoded_id).first()
-    db.session.delete(event)
-    db.session.commit()
 
-    return render_template("delete_event.html")
+    list_of_events = Event.query.all()
+    print(list_of_events)
+
+    if request.method == "POST":
+        event_name= request.form.get('event_name')
+        print(event_name)
+        event = Event.query.filter_by(name_event=event_name).first()
+        db.session.delete(event)
+        db.session.commit()
+        #list_of_events = Event.query.all()
+        return redirect(url_for('main.delete_event'))
+
+
+    return render_template("delete_event.html",events=list_of_events)
 
 
 @main.route('/')
@@ -47,25 +58,5 @@ def show_events():
     list_of_events = Event.query.all()
     print(list_of_events)
 
-    # list_of_events = [
-    #     {
-    #         "name": "cina cu Elena",
-    #         "date": "2025-01-21",
-    #         "time": "11:50",
-    #         "description": "in Bucuresti, la restaurant Vatra"
-    #     },
-    # {
-    #     "name": "pranz cu Andrei",
-    #     "date": "2025-01-25",
-    #     "time": "13:30",
-    #     "description": "la restaurant Casa Verona"
-    # },
-    # {
-    #     "name": "intalnire de afaceri",
-    #     "date": "2025-01-30",
-    #     "time": "10:00",
-    #     "description": "in Bucuresti, biroul principal"
-    # }
-    # ]
 
     return render_template("main_view.html",  events=list_of_events)
